handle delete warehouse alert

+++ app/Http/Controllers/Settings/WarehousesController.php
@@ -3,12 +3,14 @@
 namespace App\Http\Controllers\Settings;
 
 use App\Http\Controllers\Controller;
+use App\Models\Adjustment;
 use App\Models\Product;
 use App\Models\ProductWarehouse;
 use App\Models\Warehouse;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Redirect;
 use Illuminate\Validation\Rule;
 
 class WarehousesController extends Controller
@@ -159,11 +161,25 @@ class WarehousesController extends Controller
     public function destroy(string $id)
     {
         $warehouses = Warehouse::where('id', $id)->first();
-        $warehouses->delete();
-        ProductWarehouse::where('warehouse_id', $id)->update([
-            'deleted_at' => Carbon::now(),
-        ]);
+        $adjustment = Adjustment::where('warehouse_id', $id)->first();
+        try {
+            DB::beginTransaction();
+            if ($warehouses && ! $adjustment) {
+                $warehouses->delete();
+                ProductWarehouse::where('warehouse_id', $id)->update([
+                    'deleted_at' => Carbon::now(),
+                ]);
 
-        return redirect()->route('settings.warehouses.index')->with('success', 'Data Warehouse deleted successfully');
+                DB::commit();
+
+                return redirect()->route('settings.warehouses.index')->with('success', 'Data Warehouse deleted successfully');
+            } else {
+                return redirect()->back()->with('errorzz', 'Warehouse cannot be deleted because it is used in adjustments !!');
+            }
+        } catch (\Throwable $th) {
+            DB::rollBack();
+
+            return redirect()->back()->with('error', 'Data Warehouse deleted failed');
+        }
     }
 }
